[HUDI-7364] Move InLineFs classes to hudi-hadoop-common module (#10599)
