+++ hudi-common/src/test/java/org/apache/hudi/common/fs/inline/InLineFSUtilsTest.java
@@ -19,6 +19,8 @@
 package org.apache.hudi.common.fs.inline;
 
 import org.apache.hudi.common.testutils.FileSystemTestUtils;
+import org.apache.hudi.hadoop.fs.inline.InLineFSUtils;
+import org.apache.hudi.hadoop.fs.inline.InLineFileSystem;
 
 import org.apache.hadoop.fs.Path;
 import org.junit.jupiter.params.ParameterizedTest;
