+++ hudi-common/src/test/java/org/apache/hudi/common/testutils/FileSystemTestUtils.java
@@ -18,11 +18,11 @@
 
 package org.apache.hudi.common.testutils;
 
-import org.apache.hudi.common.fs.inline.InLineFSUtils;
-import org.apache.hudi.common.fs.inline.InLineFileSystem;
-import org.apache.hudi.common.fs.inline.InMemoryFileSystem;
 import org.apache.hudi.common.table.log.TestLogReaderUtils;
 import org.apache.hudi.common.util.FileIOUtils;
+import org.apache.hudi.hadoop.fs.inline.InLineFSUtils;
+import org.apache.hudi.hadoop.fs.inline.InLineFileSystem;
+import org.apache.hudi.hadoop.fs.inline.InMemoryFileSystem;
 
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
